requirements.txt
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
cachetools==5.5.0
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clip==0.2.0
comm==0.2.2
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
datasets==2.13.1
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
easydict==1.13
entrypoints==0.3
et-xmlfile==1.1.0
fanogan==0.0.1
fastjsonschema==2.20.0
flake8==3.7.9
flatbuffers==24.3.25
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
ftfy==6.2.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grad-cam==1.5.2
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.23.3
imageio==2.34.1
imgaug==0.4.0
imutils==0.5.4
ipympl==0.9.4
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.5.0
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
libclang==18.1.1
littleutils==0.2.4
Markdown==3.6
markdown-it-py==3.0.0
matplotlib==3.9.2
mccabe==0.6.1
mdurl==0.1.2
MedPy==0.5.1
mistune==3.0.2
mkl-service==2.4.0
ml-dtypes==0.4.0
monai==1.3.0
multidict==6.1.0
multiprocess==0.70.14
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.2.1
nibabel==5.2.1
notebook_shim==0.2.4
numpy==2.1.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
ogb==1.3.6
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
outdated==0.2.2
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
pillow==10.2.0
prometheus_client==0.21.0
protobuf==4.25.3
pyarrow==17.0.0
pycodestyle==2.5.0
pycparser==2.22
pydicom==2.4.4
pyflakes==2.1.1
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-json-logger==2.0.7
pytz==2024.1
referencing==0.35.1
regex==2024.5.15
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.21.0
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.14.1
segment_anything==1.0
Send2Trash==1.8.3
sentry-sdk==2.13.0
setproctitle==1.3.3
shapely==2.0.4
SimpleITK==2.3.1
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sympy==1.12.1
tabulate==0.9.0
tb-nightly==2.18.0a20240708
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.17.0
threadpoolctl==3.5.0
tifffile==2023.9.26
tinycss2==1.3.0
tokenizers==0.13.3
tomli==2.1.0
torch==2.3.1
torch-tb-profiler==0.4.3
torchaudio==2.3.1+cu118
torchvision==0.18.1
torchvision4ad==0.1.2
tornado==6.4.1
tqdm==4.66.4
transformers==4.30.2
triton==2.3.1
ttach==0.0.3
types-python-dateutil==2.9.0.20241003
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wandb==0.17.8
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.13
wilds==1.2.2
wrapt==1.16.0
xxhash==3.5.0
yacs==0.1.8
yapf==0.29.0
yarl==1.11.1
